requirements.txt
fastapi==0.111.0
uvicorn==0.32.0
pydantic==2.10.6
qdrant-client==1.10.0
elasticsearch==8.10.1
redis==5.0.1
sentence-transformers==2.2.2
langchain==0.0.335
unstructured==0.10.30
python-multipart==0.0.6
arq==0.25.0
prometheus-client==0.17.1
rank-bm25==0.2.2
python-dotenv==1.0.0
httpx==0.25.1
asyncio==3.4.3